Move the configuration parsing into it's own module without importing anything from pydoctor.
