+++ pydoctor/utils.py
@@ -1,8 +1,9 @@
 """General purpose utility functions."""
 from pathlib import Path
 import sys
-import functools
-from typing import Any, Type, TypeVar, Tuple, Union, cast, TYPE_CHECKING
+from typing import Type, TypeVar, Tuple, Union, cast, TYPE_CHECKING
+
+from pydoctor._configparser import partialclass
 
 if TYPE_CHECKING:
     from pydoctor import model
@@ -83,14 +84,4 @@ def parse_privacy_tuple(value:str, opt: str) -> Tuple['model.PrivacyClass', str]
     else:
         return (priv, parts[1].strip())
 
-def partialclass(cls: Type[Any], *args: Any, **kwds: Any) -> Type[Any]:
-    """
-    Bind a class to be created with some predefined __init__ arguments.
-    """
-    # mypy gets errors: - Variable "cls" is not valid as a type
-    #                   - Invalid base class "cls" 
-    class NewCls(cls): #type: ignore
-        __init__ = functools.partialmethod(cls.__init__, *args, **kwds) #type: ignore
-        __class__ = cls
-    assert isinstance(NewCls, type)
-    return NewCls
+partialclass = partialclass
